+++ packages/react/src/index.ts
@@ -1,6 +1,4 @@
 export * from './use-chat';
-export * from './use-chat-2';
-export { Chat2 } from './react-chat';
+export { Chat } from './react-chat';
 export * from './use-completion';
 export * from './use-object';
-export { createChatStore } from './chat-store';
