+++ packages/react/src/use-chat.ts
@@ -1,19 +1,18 @@
 import {
-  ChatStore,
+  AbstractChat,
+  AbstractChatInit,
+  ChatEvent,
   convertFileListToFileUIParts,
-  defaultChatStoreOptions,
-  generateId as generateIdFunc,
   InferUIDataParts,
   UIDataPartSchemas,
   type ChatRequestOptions,
-  type ChatStoreEvent,
   type CreateUIMessage,
   type FileUIPart,
   type UIMessage,
   type UseChatOptions,
 } from 'ai';
 import { useCallback, useRef, useState, useSyncExternalStore } from 'react';
-import { createChatStore } from './chat-store';
+import { Chat, ChatInit } from './react-chat';
 import { throttle } from './throttle';
 
 export type { CreateUIMessage, UIMessage, UseChatOptions };
@@ -25,60 +24,7 @@ export type UseChatHelpers<
   /**
    * The id of the chat.
    */
-  readonly chatId: string;
-
-  /**
-   * Hook status:
-   *
-   * - `submitted`: The message has been sent to the API and we're awaiting the start of the response stream.
-   * - `streaming`: The response is actively streaming in from the API, receiving chunks of data.
-   * - `ready`: The full response has been received and processed; a new user message can be submitted.
-   * - `error`: An error occurred during the API request, preventing successful completion.
-   */
-  readonly status: 'submitted' | 'streaming' | 'ready' | 'error';
-
-  /** Current messages in the chat */
-  readonly messages: UIMessage<
-    MESSAGE_METADATA,
-    InferUIDataParts<DATA_PART_SCHEMAS>
-  >[];
-
-  /** The error object of the API request */
-  readonly error: undefined | Error;
-
-  /**
-   * Append a user message to the chat list. This triggers the API call to fetch
-   * the assistant's response.
-   *
-   * @param message The message to append
-   * @param options Additional options to pass to the API call
-   */
-  append: (
-    message: CreateUIMessage<
-      MESSAGE_METADATA,
-      InferUIDataParts<DATA_PART_SCHEMAS>
-    >,
-    options?: ChatRequestOptions,
-  ) => Promise<void>;
-
-  /**
-   * Reload the last AI chat response for the given chat history. If the last
-   * message isn't from the assistant, it will request the API to generate a
-   * new response.
-   */
-  reload: (
-    chatRequestOptions?: ChatRequestOptions,
-  ) => Promise<string | null | undefined>;
-
-  /**
-   * Abort the current request immediately, keep the generated tokens if any.
-   */
-  stop: () => void;
-
-  /**
-   * Resume an ongoing chat generation stream. This does not resume an aborted generation.
-   */
-  experimental_resume: () => void;
+  readonly id: string;
 
   /**
    * Update the `messages` state locally. This is useful when you want to
@@ -120,58 +66,50 @@ export type UseChatHelpers<
     },
   ) => void;
 
-  addToolResult: ({
-    toolCallId,
-    result,
-  }: {
-    toolCallId: string;
-    result: any;
-  }) => void;
-};
+  error: Error | undefined;
+} & Pick<
+  AbstractChat<MESSAGE_METADATA, DATA_PART_SCHEMAS>,
+  | 'append'
+  | 'reload'
+  | 'stop'
+  | 'experimental_resume'
+  | 'addToolResult'
+  | 'status'
+  | 'messages'
+>;
+
+export type UseChatOptions2<
+  MESSAGE_METADATA = unknown,
+  DATA_TYPE_SCHEMAS extends UIDataPartSchemas = UIDataPartSchemas,
+> = (
+  | { chat: Chat<MESSAGE_METADATA, DATA_TYPE_SCHEMAS> }
+  | ChatInit<MESSAGE_METADATA, DATA_TYPE_SCHEMAS>
+) & {
+  /**
+  /**
+   * Initial input of the chat.
+   */
+  initialInput?: string;
+} & Pick<
+    AbstractChatInit<MESSAGE_METADATA, DATA_TYPE_SCHEMAS>,
+    'onToolCall' | 'onFinish' | 'onError'
+  >;
 
 export function useChat<
   MESSAGE_METADATA = unknown,
   DATA_PART_SCHEMAS extends UIDataPartSchemas = UIDataPartSchemas,
 >({
-  chatId,
   initialInput = '',
-  onToolCall,
-  onFinish,
-  onError,
-  generateId = generateIdFunc,
   experimental_throttle: throttleWaitMs,
-  chatStore: chatStoreArg,
-}: UseChatOptions<MESSAGE_METADATA, DATA_PART_SCHEMAS> & {
+  ...options
+}: UseChatOptions2<MESSAGE_METADATA, DATA_PART_SCHEMAS> & {
   /**
 Custom throttle wait in ms for the chat messages and data updates.
 Default is undefined, which disables throttling.
    */
   experimental_throttle?: number;
-} = {}): UseChatHelpers<MESSAGE_METADATA, DATA_PART_SCHEMAS> {
-  // Generate ID once, store in state for stability across re-renders
-  const [hookId] = useState(generateId);
-
-  // Use the caller-supplied ID if available; otherwise, fall back to our stable ID
-  const stableChatId = chatId ?? hookId;
-
-  // chat store setup
-  const chatStore = useRef(
-    chatStoreArg == null
-      ? createChatStore(
-          defaultChatStoreOptions<MESSAGE_METADATA, DATA_PART_SCHEMAS>({
-            api: '/api/chat',
-            generateId,
-          })(),
-        )
-      : typeof chatStoreArg === 'function'
-        ? createChatStore(chatStoreArg())
-        : chatStoreArg,
-  );
-
-  // ensure the chat is in the store
-  if (!chatStore.current.hasChat(stableChatId)) {
-    chatStore.current.addChat(stableChatId, []);
-  }
+}): UseChatHelpers<MESSAGE_METADATA, DATA_PART_SCHEMAS> {
+  const chatRef = useRef('chat' in options ? options.chat : new Chat(options));
 
   const subscribe = useCallback(
     ({
@@ -179,52 +117,34 @@ Default is undefined, which disables throttling.
       eventType,
     }: {
       onStoreChange: () => void;
-      eventType: ChatStoreEvent['type'];
-    }) => {
-      return chatStore.current.subscribe({
-        onChatChanged: event => {
-          if (event.chatId !== stableChatId || event.type !== eventType) return;
+      eventType: ChatEvent['type'];
+    }) =>
+      chatRef.current.subscribe({
+        onChange: event => {
+          if (event.type !== eventType) return;
           onStoreChange();
         },
-      });
-    },
-    [chatStore, stableChatId],
+      }),
+    [chatRef],
   );
 
   const addToolResult = useCallback(
     (
-      options: Omit<
-        Parameters<
-          ChatStore<MESSAGE_METADATA, DATA_PART_SCHEMAS>['addToolResult']
-        >[0],
-        'chatId'
-      >,
-    ) => chatStore.current.addToolResult({ chatId: stableChatId, ...options }),
-    [chatStore, stableChatId],
-  );
-
-  const stopStream = useCallback(() => {
-    chatStore.current.stopStream({ chatId: stableChatId });
-  }, [chatStore, stableChatId]);
-
-  const error = useSyncExternalStore(
-    callback =>
-      subscribe({
-        onStoreChange: callback,
-        eventType: 'chat-status-changed',
-      }),
-    () => chatStore.current.getError(stableChatId),
-    () => chatStore.current.getError(stableChatId),
+      options: Parameters<
+        Chat<MESSAGE_METADATA, DATA_PART_SCHEMAS>['addToolResult']
+      >[0],
+    ) => chatRef.current.addToolResult(options),
+    [chatRef],
   );
 
   const status = useSyncExternalStore(
     callback =>
       subscribe({
         onStoreChange: callback,
-        eventType: 'chat-status-changed',
+        eventType: 'status-changed',
       }),
-    () => chatStore.current.getStatus(stableChatId),
-    () => chatStore.current.getStatus(stableChatId),
+    () => chatRef.current.status,
+    () => chatRef.current.status,
   );
 
   const subscribeToChatStoreForMessages = useCallback(
@@ -233,7 +153,7 @@ Default is undefined, which disables throttling.
         onStoreChange: throttleWaitMs
           ? throttle(callback, throttleWaitMs)
           : callback,
-        eventType: 'chat-messages-changed',
+        eventType: 'messages-changed',
       });
     },
     [subscribe, throttleWaitMs],
@@ -241,53 +161,8 @@ Default is undefined, which disables throttling.
 
   const messages = useSyncExternalStore(
     callback => subscribeToChatStoreForMessages(callback),
-    () => chatStore.current.getMessages(stableChatId),
-    () => chatStore.current.getMessages(stableChatId),
-  );
-
-  const append = useCallback(
-    (
-      message: CreateUIMessage<
-        MESSAGE_METADATA,
-        InferUIDataParts<DATA_PART_SCHEMAS>
-      >,
-      { headers, body }: ChatRequestOptions = {},
-    ) =>
-      chatStore.current.submitMessage({
-        chatId: stableChatId,
-        message,
-        headers,
-        body,
-        onError,
-        onToolCall,
-        onFinish,
-      }),
-    [chatStore, stableChatId, onError, onToolCall, onFinish],
-  );
-
-  const reload = useCallback(
-    async ({ headers, body }: ChatRequestOptions = {}) =>
-      chatStore.current.resubmitLastUserMessage({
-        chatId: stableChatId,
-        headers,
-        body,
-        onError,
-        onToolCall,
-        onFinish,
-      }),
-    [chatStore, stableChatId, onError, onToolCall, onFinish],
-  );
-  const stop = useCallback(() => stopStream(), [stopStream]);
-
-  const experimental_resume = useCallback(
-    async () =>
-      chatStore.current.resumeStream({
-        chatId: stableChatId,
-        onError,
-        onToolCall,
-        onFinish,
-      }),
-    [chatStore, stableChatId, onError, onToolCall, onFinish],
+    () => chatRef.current.messages,
+    () => chatRef.current.messages,
   );
 
   const setMessages = useCallback(
@@ -308,12 +183,9 @@ Default is undefined, which disables throttling.
         messagesParam = messagesParam(messages);
       }
 
-      chatStore.current.setMessages({
-        id: stableChatId,
-        messages: messagesParam,
-      });
+      chatRef.current.messages = messagesParam;
     },
-    [stableChatId, messages],
+    [chatRef, messages],
   );
 
   // Input state and handlers.
@@ -334,9 +206,9 @@ Default is undefined, which disables throttling.
 
       if (!input && fileParts.length === 0) return;
 
-      append(
+      chatRef.current.append(
         {
-          id: generateId(),
+          id: chatRef.current.generateId(),
           role: 'user',
           metadata: undefined,
           parts: [...fileParts, { type: 'text', text: input }],
@@ -349,7 +221,7 @@ Default is undefined, which disables throttling.
 
       setInput('');
     },
-    [input, generateId, append],
+    [input, chatRef],
   );
 
   const handleInputChange = (e: any) => {
@@ -358,13 +230,13 @@ Default is undefined, which disables throttling.
 
   return {
     messages,
-    chatId: stableChatId,
+    id: chatRef.current.id,
     setMessages,
-    error,
-    append,
-    reload,
-    stop,
-    experimental_resume,
+    append: chatRef.current.append,
+    reload: chatRef.current.reload,
+    stop: chatRef.current.stop,
+    error: chatRef.current.error,
+    experimental_resume: chatRef.current.experimental_resume,
     input,
     setInput,
     handleInputChange,
